+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- cli: Always show percent complete if it contains a value
+
 ## [0.9.0] - 2025-09-24
 
 ### Added